Other player's pokemon party now gets sent over the network.
Scrapped old method of sending over only a small amount of data to
recreate the pokemon on the client side. I did that because I'm not sure
if I was able to fully recreate the pokemon on the client side. Now, I
just ship the entire Pokemon struct over since it has no padding and
it's easier. This ensures that I don't have any missing data for the
pokemon. This does increase the packet size by 300 bytes, but that
doesn't really seem to be a problem so far.

Moving forward, I'm going to work on the battle system and hopefully
have a playable beta version.